Create messages folder and add tools_definition module
